backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
dnspython==2.8.0
docopt==0.6.2
dotenv==0.9.9
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.129.0
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
pipreqs==0.4.13
psycopg2-binary==2.9.9
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
yarg==0.1.10
reportlab==4.2.5
pdfplumber==0.11.4
python-docx==1.1.2
groq>=1.0.0
beautifulsoup4==4.12.2
